fixed import on Home/Home.test and changed Sidebarmenu and sidemenubar names because they were similar and confusing
