+++ src/client/components/HomeLeftBar/HomeLeftBar.stories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import SidebarMenu from './HomeLeftBar.component';
+
+export default {
+  title: 'HomeLeftBar',
+  component: SidebarMenu,
+};
+
+export const sidebar = () => {
+  return <SidebarMenu />;
+};
