+++ src/client/containers/Home/Home.test.js
@@ -3,7 +3,7 @@ import '@testing-library/jest-dom/extend-expect';
 import { render } from '@testing-library/react';
 import React from 'react';
 
-import { Home } from './Home';
+import { Home } from './Home.container';
 
 test('Home section has correct classname', () => {
   const title = 'home';
